pyproject.toml
[tool.poetry]
name = "team-projector"
version = "0.1.0"
description = ""
authors = ["junte"]

[tool.poetry.dependencies]
python = "^3.8"
python-gitlab = "*"
python-dateutil = "*"
python-decouple = "*"
Django = "3.*"
djangorestframework = "*"
django-cors-headers = "*"
django-split-settings = "*"
django-extensions = "*"
django-filter = "*"
django-model-utils = "*"
django-health-check = "*"
django-activity-stream = {git = "https://github.com/justquick/django-activity-stream.git", rev = "master"}
psycopg2-binary = "*"
jnt-django-bitfield = "*"
jnt-django-admin-tools = "*"
social-auth-app-django = "*"
social-auth-core = "*"
celery = "*"
flower = "*"
redis = "3.3.11"
xlsxwriter = "*"
graphene-django = "*"
slackclient = "*"
uwsgi = "*"
uwsgitop = "*"
tqdm = "*"
black = "^19.10b0"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
pytest-xdist = "*"
factory-boy = "*"
freezegun = "*"
flake8 = "*"
flake8-django = "*"
flake8-pytest-style = "*"
flake8-pytest = "*"
mypy = "*"
httpretty = "*"
wemake-python-styleguide = "*"
polint = "*"
dennis = "*"
safety = "*"
xenon = "*"

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
    | \.git          # root of the project
    | \.mypy_cache
    | \.pytest_cache
    | .*\/migrations
  )/
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
